refactor code for better testing & benchmarking
