+++ cmd/leetcodes/timers/TimeAllFunctions.go
@@ -1,0 +1,7 @@
+package timers
+
+func TimeAllFunctions(n, iterations int) {
+	TimeContainsDuplicate(n, iterations)
+	TimeIsAnagram(n, iterations)
+	TimeTwoSum(n, iterations)
+}
